refactor: removes ops.rs and move trait impls into sep modules
